reduced memory usage of AStar algorithm (now it can solve the first 1000 test cases of codegolf26232, in spite of test case 934 being very challenging)

+++ src/colorfill/solver/AStarPuchertStrategy.java
@@ -30,13 +30,13 @@ public class AStarPuchertStrategy implements AStarStrategy {
 
     private final ColorAreaSet visited;
     private ColorAreaSet current, next;
-    private final int[] numCaNotFilled;
+    private final short[] numCaNotFilled;
 
     AStarPuchertStrategy(final Board board) {
         this.visited = new ColorAreaSet(board);
         this.current = new ColorAreaSet(board);
         this.next = new ColorAreaSet(board);
-        this.numCaNotFilled = new int[board.getNumColors()];
+        this.numCaNotFilled = new short[board.getNumColors()];
     }
 
     /* (non-Javadoc)
